requirements.txt
telethon>=1.35.0
anthropic>=0.18.0
openai>=1.0.0
aiosqlite>=0.19.0
pyyaml>=6.0.0
typer>=0.9.0
rich>=13.0.0
psutil>=5.9.0
questionary>=2.0.0
